tokens refactor - created tokens directory within scripts directory to hold all token related helper functions
